+++ src/test/kotlin/db/tables/StargazersTest.kt
@@ -7,6 +7,8 @@ import com.neelkamath.omniChatBackend.db.awaitBrokering
 import com.neelkamath.omniChatBackend.db.messagesNotifier
 import com.neelkamath.omniChatBackend.graphql.routing.PageInfo
 import com.neelkamath.omniChatBackend.graphql.routing.UnstarredChat
+import com.neelkamath.omniChatBackend.linkedHashSetOf
+import com.neelkamath.omniChatBackend.toLinkedHashSet
 import io.reactivex.rxjava3.subscribers.TestSubscriber
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Nested
@@ -83,7 +85,7 @@ class StargazersTest {
     }
 
     @Nested
-    inner class Read {
+    inner class ReadMessageIdList {
         @Test
         fun `Only the user's messages must be retrieved`() {
             val (user1Id, user2Id) = createVerifiedUsers(2).map { it.info.id }
@@ -91,13 +93,13 @@ class StargazersTest {
             val (messageId) = listOf(user1Id, user2Id).map { userId ->
                 Messages.message(userId, chatId).also { Stargazers.create(userId, it) }
             }
-            assertEquals(listOf(messageId), Stargazers.read(user1Id).edges.map { it.node.messageId })
+            assertEquals(linkedHashSetOf(messageId), Stargazers.readMessageIdList(user1Id))
         }
 
         @Test
         fun `Every item must be retrieved if neither cursor nor limit get supplied`() {
             val (adminId, messageIdList) = createStarredChat()
-            assertEquals(messageIdList, Stargazers.read(adminId).edges.map { it.node.messageId })
+            assertEquals(messageIdList.toLinkedHashSet(), Stargazers.readMessageIdList(adminId))
         }
 
         @Test
@@ -106,8 +108,8 @@ class StargazersTest {
             val index = 5
             val messageId = messageIdList[index]
             Messages.delete(messageId)
-            val actual = Stargazers.read(adminId, ForwardPagination(after = messageId)).edges.map { it.node.messageId }
-            assertEquals(messageIdList.drop(index + 1), actual)
+            val actual = Stargazers.readMessageIdList(adminId, ForwardPagination(after = messageId))
+            assertEquals(messageIdList.drop(index + 1).toLinkedHashSet(), actual)
         }
 
         @Test
@@ -115,9 +117,9 @@ class StargazersTest {
             val (adminId, messageIdList) = createStarredChat()
             val first = 3
             val index = 5
-            val expected = messageIdList.slice(index + 1..index + first)
+            val expected = messageIdList.slice(index + 1..index + first).toLinkedHashSet()
             val pagination = ForwardPagination(first, after = messageIdList[index])
-            val actual = Stargazers.read(adminId, pagination).edges.map { it.node.messageId }
+            val actual = Stargazers.readMessageIdList(adminId, pagination)
             assertEquals(expected, actual)
         }
 
@@ -125,15 +127,16 @@ class StargazersTest {
         fun `The number of items specified by the limit from the first item must be retrieved when there's no cursor`() {
             val (adminId, messageIdList) = createStarredChat()
             val first = 3
-            val actual = Stargazers.read(adminId, ForwardPagination(first)).edges.map { it.node.messageId }
-            assertEquals(messageIdList.take(first), actual)
+            val actual = Stargazers.readMessageIdList(adminId, ForwardPagination(first))
+            assertEquals(messageIdList.take(first).toLinkedHashSet(), actual)
         }
 
         @Test
         fun `When requesting items after the start cursor, 'hasNextPage' must be 'false', and 'hasPreviousPage' must be 'true'`() {
             val (adminId, messageIdList) = createStarredChat()
-            val (hasNextPage, hasPreviousPage) =
-                Stargazers.read(adminId, ForwardPagination(after = messageIdList[0])).pageInfo
+            val pagination = ForwardPagination(after = messageIdList[0])
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            val (hasNextPage, hasPreviousPage) = Stargazers.readPageInfo(adminId, lastMessageId, pagination)
             assertFalse(hasNextPage)
             assertTrue(hasPreviousPage)
         }
@@ -143,25 +146,29 @@ class StargazersTest {
             val (adminId, messageIdList) = createStarredChat()
             val index = 5
             val pagination = ForwardPagination(after = messageIdList[index])
-            val actual = Stargazers.read(adminId, pagination).edges.map { it.node.messageId }
-            assertEquals(messageIdList.drop(index + 1), actual)
+            val expected = messageIdList.drop(index + 1).toLinkedHashSet()
+            val actual = Stargazers.readMessageIdList(adminId, pagination)
+            assertEquals(expected, actual)
         }
 
         @Test
         fun `Given items 1-10 where item 4 has been deleted, when requesting the first three items after item 2, then items 3, 5, and 6 must be retrieved`() {
             val (adminId, messageIdList) = createStarredChat()
             Messages.delete(messageIdList[3])
-            val edges = Stargazers.read(adminId, ForwardPagination(first = 3, after = messageIdList[1])).edges
-            assertEquals(listOf(messageIdList[2], messageIdList[4], messageIdList[5]), edges.map { it.node.messageId })
+            val paginatedMessages =
+                Stargazers.readMessageIdList(adminId, ForwardPagination(first = 3, after = messageIdList[1]))
+            assertEquals(linkedHashSetOf(messageIdList[2], messageIdList[4], messageIdList[5]), paginatedMessages)
         }
     }
 
     @Nested
-    inner class BuildPageInfo {
+    inner class ReadPageInfo {
         @Test
         fun `When requesting zero items sans cursor, the page info must indicate such`() {
             val (adminId) = createStarredChat()
-            val (hasNextPage, hasPreviousPage) = Stargazers.read(adminId, ForwardPagination(first = 0)).pageInfo
+            val pagination = ForwardPagination(first = 0)
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            val (hasNextPage, hasPreviousPage) = Stargazers.readPageInfo(adminId, lastMessageId, pagination)
             assertTrue(hasNextPage)
             assertFalse(hasPreviousPage)
         }
@@ -170,7 +177,8 @@ class StargazersTest {
         fun `When requesting zero items after the end cursor, the 'hasNextPage' and 'hasPreviousPage' must indicate such`() {
             val (adminId, messageIdList) = createStarredChat()
             val pagination = ForwardPagination(first = 0, after = messageIdList.last())
-            val (hasNextPage, hasPreviousPage) = Stargazers.read(adminId, pagination).pageInfo
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            val (hasNextPage, hasPreviousPage) = Stargazers.readPageInfo(adminId, lastMessageId, pagination)
             assertFalse(hasNextPage)
             assertTrue(hasPreviousPage)
         }
@@ -179,7 +187,8 @@ class StargazersTest {
         fun `Given items 1-10, when requesting zero items after item 5, the 'hasNextPage' and 'hasPreviousPage' must indicate such`() {
             val (adminId, messageIdList) = createStarredChat()
             val pagination = ForwardPagination(first = 0, after = messageIdList[4])
-            val (hasNextPage, hasPreviousPage) = Stargazers.read(adminId, pagination).pageInfo
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            val (hasNextPage, hasPreviousPage) = Stargazers.readPageInfo(adminId, lastMessageId, pagination)
             assertTrue(hasNextPage)
             assertTrue(hasPreviousPage)
         }
@@ -188,31 +197,35 @@ class StargazersTest {
         fun `Zero items must be retrieved along with the correct 'hasNextPage' and 'hasPreviousPage' when using the last item's cursor`() {
             val (adminId, messageIdList) = createStarredChat()
             val pagination = ForwardPagination(after = messageIdList.last())
-            val (edges, pageInfo) = Stargazers.read(adminId, pagination)
-            assertEquals(0, edges.size)
-            assertEquals(false, pageInfo.hasNextPage)
-            assertEquals(true, pageInfo.hasPreviousPage)
+            val paginatedMessages = Stargazers.readMessageIdList(adminId, pagination)
+            val (hasNextPage, hasPreviousPage) =
+                Stargazers.readPageInfo(adminId, paginatedMessages.lastOrNull(), pagination)
+            assertEquals(0, paginatedMessages.size)
+            assertEquals(false, hasNextPage)
+            assertEquals(true, hasPreviousPage)
         }
 
         @Test
         fun `Retrieving the first of many items must cause the page info to state there are only items after it`() {
             val (adminId, messageIdList) = createStarredChat()
-            val pageInfo = Stargazers.read(adminId, ForwardPagination(first = 1, after = messageIdList[0])).pageInfo
-            assertTrue(pageInfo.hasNextPage)
+            val pagination = ForwardPagination(first = 1, after = messageIdList[0])
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            Stargazers.readPageInfo(adminId, lastMessageId, pagination).hasNextPage.let(::assertTrue)
         }
 
         @Test
         fun `Retrieving the last of many items must cause the page info to state there are only items before it`() {
             val (adminId, messageIdList) = createStarredChat()
             val pagination = ForwardPagination(after = messageIdList[messageIdList.size - 2])
-            Stargazers.read(adminId, pagination).pageInfo.hasPreviousPage.let(::assertTrue)
+            val lastMessageId = Stargazers.readMessageIdList(adminId, pagination).lastOrNull()
+            Stargazers.readPageInfo(adminId, lastMessageId, pagination).hasPreviousPage.let(::assertTrue)
         }
 
         @Test
         fun `If there are zero items, the page info must indicate such`() {
             val adminId = createVerifiedUsers(1).first().info.id
             val expected = PageInfo(hasNextPage = false, hasPreviousPage = false, startCursor = null, endCursor = null)
-            assertEquals(expected, Stargazers.read(adminId).pageInfo)
+            assertEquals(expected, Stargazers.readPageInfo(adminId, lastMessageId = null))
         }
 
         @Test
@@ -224,13 +237,14 @@ class StargazersTest {
                 startCursor = messageIdList[0],
                 endCursor = messageIdList[0],
             )
-            assertEquals(expected, Stargazers.read(adminId).pageInfo)
+            val actual = Stargazers.readPageInfo(adminId, lastMessageId = messageIdList[0])
+            assertEquals(expected, actual)
         }
 
         @Test
         fun `The first and last cursors must be the first and last items respectively`() {
             val (adminId, messageIdList) = createStarredChat()
-            val (_, _, startCursor, endCursor) = Stargazers.read(adminId).pageInfo
+            val (_, _, startCursor, endCursor) = Stargazers.readPageInfo(adminId, lastMessageId = null)
             assertEquals(messageIdList[0], startCursor)
             assertEquals(messageIdList.last(), endCursor)
         }
@@ -245,7 +259,7 @@ class StargazersTest {
             val messageId = Messages.message(adminId, chatId)
             Stargazers.create(userId, messageId)
             GroupChatUsers.removeUsers(chatId, userId)
-            assertTrue(Stargazers.read(userId).edges.isEmpty())
+            assertTrue(Stargazers.readMessageIdList(userId).isEmpty())
         }
 
         @Test